pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.66.0",
    "deepgram-sdk>=4.8.1",
    "faster-whisper>=1.1.1",
    "google-genai>=1.33.0",
    "gtts>=2.5.4",
    "httpx>=0.28.1",
    "keyboard>=0.13.5",
    "mcp[cli]>=1.13.1",
    "numpy>=2.2.6",
    "openai-whisper>=20250625",
    "protobuf>=6.32.0",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "pynput>=1.8.1",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.1.1",
    "realtimestt>=0.3.104",
    "scipy>=1.15.2",
    "simpleaudio>=1.0.4",
    "sounddevice>=0.5.2",
    "speechrecognition>=3.14.3",
    "spotipy>=2.25.1",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchvision>=0.23.0",
    "whisper>=1.1.10",
]



# Specific versions of torch, torchvision, and torchaudio are installed based on the extra feature
[project.optional-dependencies]
cpu = ["torch==2.8.0", "torchvision==0.23.0", "torchaudio==2.8.0"]
cu126 = ["torch==2.8.0", "torchvision==0.23.0", "torchaudio==2.8.0"]


[tool.uv]
# The cpu and cu124 extras are mutually exclusive, you can't use both at the same time
conflicts = [[{ extra = "cpu" }, { extra = "cu126" }]]

# The package is only installed on Linux systems
# environments = ["sys_platform == 'linux' and os_name == 'posix'"]

[tool.uv.sources]
# The torch, torchvision, and torchaudio packages are installed from the pytorch-cpu and pytorch-cu124 indexes
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu126", extra = "cu126" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu126", extra = "cu126" },
]
torchaudio = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu126", extra = "cu126" },
]


[[tool.uv.index]]
# The pytorch-cpu index is used to install the torch, torchvision, and torchaudio packages
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
